requirements.txt
streamlit
scikit-learn
webcolors
rembg
backgroundremover
opencv-python
pillow
requests
numpy
colorthief